refactor: extract error page to shared template module
Creates a new `error_pages` module that consolidates:
- Shared CSS styles for error pages (dark theme, consistent typography)
- `render_error_page()` - template render error display
- `render_404_page()` - page not found display
- `ansi_to_html()` - ANSI escape code to HTML conversion utility
- `RENDER_ERROR_MARKER` - marker for detecting render errors in builds

Benefits:
- Single source of truth for error page styling
- Shared styles between template errors and 404 pages
- ANSI-to-HTML conversion is now a reusable utility
- Cleaner separation of concerns (HTML generation vs rendering logic)
- Includes test coverage for ANSI conversion and page rendering

Fixes #11
